Rewrite connection forwarding API
This commit introduces new classes to support setting up listening
sockets and port forwarding which support fully asynchronous DNS
lookups and other operations when being set up. The previous APIs
required that the setup be completed and a success or failure
indication be returned synchronously when the listener was opened.

This commit also bumps the version number to 0.5.0 and updates the
change log.