Convert from asyncore to asyncio, and redesign API to match
This commit is a major revamp of the code to migrate from the
Python asyncore framework to the new Python 3.4 asyncio framework.
All the APIs have been adapted to fit the new asyncio paradigm,
using coroutines wherever possible to avoid the need for
callbacks when performing asynchronous operations.

This commit also bumps the version number to 0.6.0 and updates the
change log.