Abstract AsyncSSH crypto API
This commit abstracts the API calls to DSA, RSA, and symmetric ciphers.
Instead of relying solely on PyCrypto for these functions, AsyncSSH can
now use either PyCA, PyCrypto, or a combination of the two to satisfy
all of its dependencies on cryptographic functions.

This commit also adds support for AES-GCM when PyCA is available, using
it as both the encryption and MAC functions when it is selected.