Fix default keyboard-interactive client authentication
This commit fixes a bug in the default handling of keyboard-interactive
authentication in SSHClientConnection. The code was intended to respond
with the password provided when the connection was first created, but
the wrong function was being called to retrieve this value.

This fix also loosens the matching done on the keyboard-interactive
prompt to allow it to work correctly when logging into Cisco devices
that use 'password: ' as a prompt instead of 'Password:'. Thanks go
out to Sam Crooks for reporting this issue and helping me to debug it.