Fix a couple of race conditions in SSHChannel
This commit fixes a couple of race conditions in SSHChannel. The first
is a crash which occurs when a channel is closed before a session has
had a chance to fully start, and the second is a problem where a
session's data_received() callback could be called before
session_started() if data arrives on a channel immediately after it is
opened.

My thanks to Michael Keller for reporting these issues and suggesting
possible fixes.