Add support for Ed25519 keys and the OpenSSH private key format
This commit adds support for Ed25519 private & public keys and the
OpenSSH private key format needed to import them. It introduces a
dependency on the Python libnacl wrapper for Ed25519 encryption, and on
the Python bcrypt package to support import and export of encrypted
OpenSSH format private keys.

This commit also adds support for import and export of DSA, RSA, and
ECDSA private keys in OpenSSH format.

Other improvements here include support for importing PKCS#8 v2 private
keys that include an embedded public key and for ECDSA keys which are
encoded using explicit ECParameters rather than a NamedCurve OID, if
the parameters match one of the supported named curves.