Add missing code to resume reading on a paused stream when needed
This commit fixes the read functions in SSHStreamSession to properly
resume reading on a stream which was previous paused when enough data
has been read to fall below the configured buffer limit.

To make things more efficient, resume_reading() function now checks
if the stream was previously paused so it can return quickly when
nothing needs to be done.

Thanks go to Michael Keller for reporting this bug and providing a fix
for it!