Integrate known hosts improvements from Alexander Travov
This commit integrates improvements proposed by Alexander Travov to
known_hosts file parsing. Specifically, it adds support for negated
patterns, hashed entries, and handling of the "@revoked" marker.

When a server offers a revoked host or CA key, the client now raises a
DisconnectError indicating that the key was revoked. Also, a
DisconnectError is raised if host key validation is enabled but no
trusted host or CA keys are found. Previously, this caused the client to
offer no host key algorithms, causing a "Connection lost" error when the
server rejected the connection.

There are a few differences in this implementation than the original. In
particular, glob patterns are supported on the full host and port when a
connection is made to a non-standard port. Also, the system will fall
back to matching on entries which do not contain a port if no keys are
found for a non-standard port. Finally, the code was refactored to do
the parsing of the known_hosts entries only once for both of these lookups
for efficiency.

Thanks Alexander for your contribution!