Support passing an explicit known_hosts file to AsyncSSH
This commit changes the create_connection and open_connection functions
to take a single "known_hosts" argument rather than separate arguments
of "server_host_keys", "server_ca_keys", and "revoked_server_keys".

As before, if none of these arguments are specified, AsyncSSH will look
for known hosts in the file ``.ssh/known_hosts``. However, it is now
possible to specify an alternate filename to use here, or a byte string
in known_hosts format.

The known_hosts argument can also contain a sequence of three values
corresponding to the list of server host keys to trust, the list of
server CA keys to trust, and the list of revoked keys to check for.
In this case, each of these values can be in any of the forms accepted
when passing public key lists into AsyncSSH.