Add IP address matching in known_hosts
This commit extends the known_hosts matching code to match on both the
hostname requested and the IP address that AsyncSSH connected to. Now
entries in known_hosts which match either the destination hostname or IP
will be considered valid when matching server host keys.

When looking up the hostname without the port, this commit also adds
the ability to match on CIDR-style network specifications, and the class
used to do this matching has been generalized to be usable later in
other places such as matching client host/IP in authorized_keys. This
support will be coming soon.