Add checks in test_keys for optional packages
This commit adds checks in test_keys for the presence of the bcrypt and
libnacl optional packages. When these packages are not installed, the
corresponding tests which rely on them are automatically skipped.

This change also temporarily comments out the GCM and Chacha ciphers in
the openssh key testing due to a bug in OpenSSH:

    https://bugzilla.mindrot.org/show_bug.cgi?id=2366

Once this bug is fixed, testing of these ciphers can be re-enabled. In
the meantime, AsyncSSH's use of these ciphers should completely work,
but OpenSSH will not be able to properly decrypt keys encrypted with
these ciphers, even when it is doing the encryption.