Add simplified asyncssh.connect() API
This commit adds a simplified asyncssh.connect() wrapper around
asyncssh.create_connection() that can be used when the caller doesn't
need to provide a customized SSHClient factory. It also updates the
majority of the client examples to use this simplified API, adds a
longer example showing custom SSHClient and SSHClientSession subclasses,
and fixes a minor issue in a couple of the streams examples.