Initial cut of SFTP client support
This commit adds a first cut of SFTP client support. It's a reasonably
complete implementation of an SFTP version 3 client, but it does not yet
support any extensions. Those will be added over time.

Some framework for adding SFTP server support is also present here and
support for that will come in future commits.

Support for later SFTP protocol versions may also be added, though it
appears that those are not yet supported by OpenSSH, which is what I'm
doing my testing against so far.

Note: This implementation attempts to follow the SFTP standard with
regard to the order of arguments for the FXP_SYMLINK message. However,
it has a workaround to automatically detect OpenSSH and paramiko peers
and reverse these arguments to be compatible with them. Other
implementations can easily be added by name if they have this problem.