Continue to enhance SFTP support
This commit adds an optional 'offset' parameter to the SFTPFile read()
and write() methods, allowing an absolute position to be specified to
start the read or write at. This is useful for callers wanting to schedule
multiple read or write operations in parallel, as there's no way to be
certain of the file position in that case.

This commit also adds a 'preserve' argument to the SFTPClient get(),
put(), and copy() methods which preserves the file permissions and last
access and modification files when uploading, downloading, or copying
files.