Add recursive upload, download, and copy support to the SFTP client
This commit adds a "recurse" option to the SFTPClient get(), put(),
and copy() methods which allows for the recursive copy of an entire
directory tree.

Also added is a "follow_symlinks" option which controls whether or not
symbolic links are followed when content is copied. If this option is
set to True, the contents of the target of the link is copied. Otherwise,
the link itself is copied. The default is to copy the link itself.

Finally, to better manage errors when doing a recursive copy, an
"error_handler" argument can be specified. It will be called with any
exceptions which occur during the copy, instead of the error being
raised. If the handler wishes to abort copying the remainder of the
files, it can raise the exception. Otherwise, the copy continues
starting at the next file.