Support multiple source files/directories in SFTP client
This commit extends the get(), put() and copy() methods of the SFTP
client to accept either a single path or a sequence of multiple paths
as the content to download, upload, or copy. When specifying multiple
paths, the destination must be an existing directory.

Also, the recusive copy support was updated to properly return an error
when trying to recursively copy a directory on top of an existing plain
file or symbolic link.