Use posixpath when operating on remote file paths
This commit switches to using posixpath rather than os.path when
operating on remote SFTP file paths, since SFTP specifies that the path
separator should always be '/'. Eventually, some of this pathname
manipulation can be done on the server for server version 6 or later,
but it will still need to be done on the client for older servers.

This also fixes a bug in the implementation of _LocalFile.close().