Add support for glob matching and glob-based download/upload/copy
This commit adds support for a match() function in SFTPClient which
matches remote files against a glob pattern. It also adds mget(),
mput(), and mcopy() methods that work like get(), put(), and copy() but
allow wildcards to be specified to determine which files should be
transferred.

This commit also fixes issues in the handling of path name encodings.
The readdir(), listdir(), and realpath() methods now all return the same
type as the path argument passed into them, even when a path encoding
is set, while getcwd() will always return a string if an encoding is
set. When no path encoding is set all methods will only accept and return
raw bytes for path names.