Flesh out the framework for adding SFTP server support
This commit begins to add the framework for full SFTP server support.
It provides a simpler way to specify handlers for new sessions when
creating an SSH server, and provides a new listen() call which can be
used to create an SSH server when a custom SSHServer instance isn't
needed.

One of the handlers which can be provided using this new mechanism is
sftp_factory, which can be given a subclass of SFTPServer that will
be used when SFTP client requests are received. If this handler is
specified, requests to start the 'sftp' subsystem will be allowed and
a new SFTPServer object will be created for each subsystem request.
This class will then be called to process the client's SFTP requests,
once this work is complete.