Add initial version of SFTP server support
This commit adds an initial version of SFTP server support for AsyncSSH.
By default, it provides access to whatever files and directories are
accessible to the user it is running as. This can be changed by passing
in a virtual "chroot" directory when an SFTP server instance is created,
though, or by implementing methods which map client requested path names
to and from local path names.

All of the SFTP methods can also easily be overriddden in the SFTPServer
class, to add whatever additional functionality the application needs.
AsyncSSH takes care of all the SFTP protocol details, leaving the
application to only implement file access methods.

This commit also cleans up a few minor documentation errors and a
missing import for SFTPVFSAttrs.