Rework chroot handling for SFTPServer symlink implementation
This commit reworks the chroot enforcement when using SFTPServer to
create a symlink with a relative target path. If the mapped version
of the target of the link points at a different path than what would be
accessed by creating a link with that relative path in the mapped
version of the new path, the relative path is adjusted to point at the
path which would have been accessed in a chroot environment.

This protection only applies to symlinks created via SFTP. If there are
existing symlinks within the tree specified as the chroot which point
elsewhere in the filesystem, SFTP will stil be able to access files
reachable through those links, and even create new symlinks to those
locations.