Update README and examples to cover SFTP support
This commit adds an entry in the README and new examples for the
recently-added SFTP client and server support. It also simplifies some
of the other examples, taking advantage of new capabilities such as the
asyncssh.listen() function and the ability to pass in a session factory
when using the streams API to avoid needing a custom subclass of
SSHServer.

All server examples have also been updated to enable authentication,
with most setting a shared authorized keys file for this purpose.