Allow SFTP server to be started from non-streams API
This commit adds the ability for applications using the non-streams
version of the API to start an SFTP server from within their
session_started() callback. This call must be made from within
session_started() before any data is read or written, and generally
should only be called when the client has requested the SFTP
subsystem on that session.