Perform pylint cleanup
This commit represents a first pass at cleaning up errors reported by
pep8 and pylint. A pylintrc is also checked in as part of this with the
configuration parameters used there. Some warnings are intentionally
suppressed for now, and will be addressed in future passes.

NOTE: Some pep8 errors will not be fixed, and unfortunately there's no
      good way to annotate the source code to flag these in the way I
      have for pylint.