Work around context manager race condition in SSHConnection
When shutting down, there seems to be a race between closing the event
loop and running the context manager exit function for SSHConnection
objects. As a result, a Runtime error of "Event loop is closed" can be
triggered.

Originally, I attempted to avoid this error by checking loop.is_closed()
in the exit function before using the event loop. However, this function
is only available in Python 3.4.2 and later, and AsyncSSH is supposed to
work with 3.4.0 and later.

For now, I've changed the SSHConnection __exit__ code to catch the
RuntimeError rather than testing is_closed(). I'll switch back to
is_closed() in the future if there are other reasons for AsyncSSH to
require a later version of Python.