Allow Unicode characters to be split across SSH data messages
This commit allows bytes which make up a single Unicode character to be
split across successive SSH data messages. In such cases, the character
will be delivered to the application as soon as all of its bytes have
been received.

As before, if the byte stream cannot be decoded using the specified
encoding or an EOF is received while in the middle of decoding a
character, a DisconnectError will be raised reporting a Unicode
decode error.