Introduce internal PacketDecodeError exception
SSH packetization is used in multiple places in the code now, and not
all packet decode errors should result in the SSH connection being
closed. For example, a decode error could occur while reading in an SSH
key or certificate in OpenSSH format.

Prior to this change, any SSH packet decode errors raised DisconnectError.
While this could be caught before doing a disconnect, it wasn't very clean.
This commit adds an internal PacketDecodeError exception which is returned
when a packet decode error occurs, and makes it the responsibility of the
calling code to catch that and translate it to the appropriate public
exception. So, packet decode errors on SSH packets received on the wire
turn into DisconnectError with code DISC_PROTOCOL_ERROR but packet
decode errors during public key import turn into KeyImportError, and
decode errors in verify public key signatures return that the signature
is not valid.