Support a broader set of open mode flags in SFTP server
This commit changes the SFTPServer open method to take the raw pflags
value from SFTP as an argument rather than a Python open mode. This is
necessary to allow the SFTP server to support a broader set of open
modes than what Python's open() function supports. Not all combinations
of flags are useful, but a number of useful combinations can't be
directly translated into standard Python open modes.

The default implemented of open() in SFTPServer has also been changed
to support these new combinations of flags by using the opener argument
in Python's open() call to pass flags directly into os.open(). The
portable SFTP flag values are converted into the local OS flag values
and then passed through.

Unfortunately, this is not a backward compatible change in the arguments
to SFTPServer.open(). So, if you have application code which expects a
Python mode as an argument, it will need to be changed to expect a
pflags value instead.

This commit also changes SFTPClient.open() to accept a pflags value, but
in this case the function was changed to allow either pflags or a Python
open mode string. So, backward compatibility is maintained as long as
the argument name was not used explicitly by the caller. For clarity,
the argument name was changed from "mode" to "pflags_or_mode".