Finish off public key unit tests
This commit adds unit tests for the SSH certificate support, and a few
other previously uncovered places in the public_key module.

To properly generate test certificates, the encode_ssh_public() method
in the SSH key classes has been refactored to be symmetric with the
decode_ssh_public() function, not automatically adding in the key
algorithm. This is now done in the calling code where it is needed.

In addition, a bug was fixed when a passphrase was passed as already
encoded bytes, rather than as a string, and unit tests were added to
cover both cases. Passing in bytes can be tricky, though, as the proper
encoding of those bytes will vary depending on the cipher and PBE
version selected.

A handful of lines were also explicitly marked for exclusion from
coverage, as they would only be executed when some of the supporting
modules are not installed.