Add unit tests for known_hosts module
This commit adds unit tests for the known_host module, and cleans up a
an unnecessary conditional check in that module. It also changes file
access to use a context manager, to make sure the file is closed
immediately after reading it.

This commit also breaks a few things off into a utility module so
multiple test modules can share them, and moves the coveragerc up a
level in the tree.