Add PyCA EC support and update AsyncSSH crypto dependencies
This commit updates AsyncSSH's dependencies to make PyCA (cryptography)
version 1.0.0 or later required and removes the older PyCrypto support.
As part of doing this, RC2 ciphers are no longer supported for
encrypting private keys, so the code associated with that has also
been removed.

Now that libnacl/libsodium is support, this commit also drops the older
curve25519-donna support. If you want curve25519 Diffie-Hellman key
exchange, you need to have libnacl and libsodium installed.

This commit also tightens up the shim code for PyCA and extends it to
support ECDSA key signing and verification. It also refactors ECDH
support so that Curve25519 & EC can share a key exchange implementation
and prepares the code for adding a PyCA shim for ECDH as soon as it
becomes available.

The DSA and RSA key exchange implementations have also been improved to
do better error checking of the key exchange response, making sure that
the signature is the right length and that there's no extra data in the
message after the signature.

The RSA implementation has also been changed to pass the optional RSA
private key parameters into PyCA, speeding up the performance of
importing RSA keys.