Fix unwanted interaction between eof_received and handling of exit status
This commit changes the code which calls eof_received() on SSHSession
objects to send an EOF rather than a full close when eof_received()
returns False, indicating that the session doesn't want the channel to
remain half-open. Sending EOF makes sure that no further data can be
sent on the channel in either direction, but it leaves open the ability
for the SSHSession to still send or receive an exit status or exit
signal before the channel is closed. Previously, the channel was
completely closed here, preventing the exit status from being delivered.

This change applies both client & server sessions receiving an EOF from
their peer. Note that the channel will still be fully closed when the
peer sends a close, though. After that, no further communication is
possible.