Move bcrypt check until it is actually needed
This commit moves the check in the private key encryption and decryption
functions for whether bcrypt is available down to just before it is
actually needed. This allows other error checking to happen first, so
better errors can be returned if a passphrase is not specified or the
cipher or KDF is unknown.