Update checks in unit tests for bcrypt and libnacl
All unit tests which depend on the installation of the bcrypt or
libnacl libraries now properly check for the presence of these libraries
before the tests are run. Previously, this was done in some places, but
the checks were missing in others. The unit tests should now run cleanly
even when these libraries are not installed.