Remove support for OpenSSH v00 certificates
This commit removes support for version "v00" OpenSSH DSA & RSA
certificates. Support for these was recently removed in the OpenSSH 7.0
release, so this support can no longer be tested with the latest version
of OpenSSH.

Looking back at the history, v00 certificates were first supported in
OpenSSH 5.4 released in March of 2010, and superseded by version v01
certificates in OpenSSH 5.6 in August of 2010, with a note that v00
certificate support would be maintained for one year and then removed.
The removal didn't happen until much later, but it seems safe to follow
OpenSSH's lead and drop support for the older version at this point.