Add login timeout to SSH server
This commit adds a new login_timeout argument to the create_server() and
listen() functions which controls the maximum time a connection is
allowed to stay open without finishing authentication. By default, this
is set to 2 minutes.

Previously, connections which didn't complete authentication were
allowed to stay open indefinitely. If this is needed for some reason,
this can be accomplished by passing in a login timeout of 0, but it
is not recommended.