Fix conversion of client keyboard-interactive to password auth
This commit fixes an issue with the automatic conversion of client
keyboard-interactive auth to password auth. With this change, the
automatic conversion to password auth will only occur when the client
doesn't override kbdint_challenge_received. Also, it will only attempt
to authenticate once, avoiding a possible loop if the server challenge
does not ask for a password.