Catch and report uncaught exceptions in SSH processing
This commit adds a new "internal_error" method on SSH connection objects
and arranges for it to be called whenever an uncaught exception occurs
during SSH processing. In this case, the exception is logged if debug
logging is enabled and the associated connection is disconnected with
a reason taken from the exception which occurred.

This change also applies to exceptions in asynchronous tasks started
during SSH processing, such as authentication handlers declared as
coroutines.