Fix a few issues related to TCP/IP forwarding listeners
This commit adds a missing response message which acknowledges requests
to cancel a TCP/IP forwarding listener. It also adds code which will
unblock callers with a failure on all global requests which haven't
received a response at the time a connection is cleaned up and makes
the clean up code more robust if an error occurs in the middle of
closing a listener.