Rework channel state machine
This commit reworks the channel state machine to perform clean up on the
channel only after a close is both sent and received and eliminates
the unnecessary 'open_sent' and 'open_received' send states. It also now
allows an EOF to be written to a channel which is being closed, avoiding
a race when both ends initiate a close at the same time.