Refactor port forwarding code
This commit refactors the code which opens a local listening socket
used for port forwarding so that it can be shared by both the
SSHClientConnection and SSHServerConnection classes.

It also fixes a race condition which could sometimes trigger an exception
if data was received on a local socket too quickly, before the channel
was set up. The code previously attempted to use pause_reading() for
this, but the asyncio socket implementation doesn't guarantee that the
pause will take effect immediately, even if it is called in the
connection_made callback.

Finally, this fixes an exception handler to look for ChannelOpenError
rather than DisconnectError when an attempt to start port forwarding
is rejected.