Add support for ssh-agent
This commit adds basic support for allowing ssh-agent to perform operations
used in public key authentication. This support is enabled automatically
if SSH_AUTH_SOCK is set in the environment and client keys are not
explicitly provided. A specific path to the ssh-agent can also be provided
as an argument when opening a new client connection.

To support this functionality, some rework was required to allow
additional methods in SSHConnection to run asynchronously, including the
code which loads client keys and the code which performs signing using
those keys.

This commit also adds a new SSHKeyPair class which bundles a private key
and its associated public data (either a key or certificate) to make it
easier to abstract when using keys stored in the agent.

This commit also adds a common facility to start async tasks in the
SSHConnection class, catching errors in those tasks and delivering them
back to the main connection object to improve error reporting.

Finally, the test framework was enhanced to better support async unit
tests, and some tests were updated to remain consistent with internal
API changes.