Rework and document SSH agent and perform some other minor cleanup
This commit reworks the recently added SSH agent support to query the
SSH client keys from an agent before attempting to open an SSH client
connection, avoid some ugly synchronization code which was previously
added. Some other minor bugs were fixed and documentation is now
available for the new SSHAgentClient class and the connect_agent
unction. Also, the README has been updated to list SSH agent support
as one of AsyncSSH's feature.

This commit also introduces a new SSHKeyPair abstract class which is
used to hold private keys and their associated public key or
certificate. Versions of this class exist for locally provided keys and
for references to keys held in an SSH agent. In the latter case, any
signing operations will automatically be performed in the agent. This
new abstraction also allows keys gotten from the SSHAgentClient's
get_keys method to be passed in via the client_keys argument to
create_connection(), supporting the ability to specify a mixture of
local keys and agent-management keys when making a connection. It also
allows for the caller to select which of an agent's keys to use by
querying them explicitly and only passing in a subset of them in the
connect call.

In addition to the above, some copyright notices were updated to
2016 and other minor errors in the documentation were fixed.