Add a wait_closed method to SSHConnection
This commit adds a new wait_closed method to SSHConnection. It is a
coroutine which can be yielded from which will block until the
SSHConnection has finished closing and its owner's connection_lost
method has been called.

This also fixes a bug where SSHConnection's _cleanup method was not
being called when a connection was closed explicitly via the close
or disconnect methods, or via the use of a context manager.