Don't call disconnect() when handling internal errors
This commit fixes a problem related to _cleanup not receiving the
right exception when an handling an internal error. A previous fix added
a missing call to _cleanup in the disconnect function, but disconnect
was being called from internal_error, causing _cleanup to now be called
with an exception of None instead of the proper exception. It was then
called again with the proper exception, but by then the cleanup had
already occurred.