Switch SFTP client and server to use streams API
This commit changes the SFTP client and server code to use the streams
API when reading and writing the SSH channel.

This commit also updates all of the places in AsyncSSH that create new
tasks to use a create_task() method in the SSHConnection class, so that
these tasks can be properly cleaned up when the connection is closed.
For this to happen, the wait_closed() coroutine on SSHConnection must be
called.