Add support for UNIX domain socket forwarding
This commit adds new methods to SSHConnection and associated support
classes to support opening direct and forwarded UNIX domain socket
connections compatible with the implementation of this in OpenSSH.

This commit also fixes the signature of open_connection() on
SSHServerConnection to not include a handler_factory argument. This
argument should only have been present on start_server and the
newly added start_unix_server.

This commit also fixes the implementation of wait_closed() on
SSHForwardListener to properly wait for the listener to fully close
if it is called after close() but the close is still in progress.

Some exception messages were also improved in this commit.