Initial cut at Windows compatibility
This commit makes some minor changes to the connection and sftp modules
to allow AsyncSSH to load and run on Windows.

Some features such as UNIX domain socket access will not work on
Windows, and additional work would be needed to allow the unit tests
to run there, but these changes should be enough to get basic
functionality working.