Avoid race in closing SSH connection
This commit avoids a race condition which would lead to an exception
when cleaning up open channels or listeners during close of an SSH
connection. The cleanup was deferred until after the connection had
already cleared its part of this state. The fix leaves the state
in the connection in place until the channels and listeners can
explicit call in to clear it.