Update channel state machine to better handle close when paused
This commit reworks the state machine for SSH channels to defer
reporting a channel close when there's still data waiting to be
delivered to the application. This can occur when the application has
called pause_reading() on the channel. Previously, if a close arrived in
this state, any data not yet delivered would be lost, or in some cases
delivery would be attempted after the channel was closed, which would
cause an exception to be raised.