Split known host parsing and matching into separate calls
This commit provides new functions to split parsing of a known hosts
file and matching against it into separate calls. This reduces the cost
of parsing in applications which make connections to several different
hosts using a single known_hosts file. Thanks go out to Josh Yudaken
for suggesting this and providing a sample implementation.