Fix error handling code when opening a local TCP listener
This commit adds a missing close call when cleaning up after a failure
to open a local TCP listening socket. Previously, the code could leak a
socket object in this case.

This commit also adds an extra guard to allow close to safely be called
multiple times on a listener and disables coverage for some rare hard
to trigger error conditions.