Don't change recv state on SSHChannel on application close/abort
A recent commit changed the abort & close calls on SSHChannel to
update the recv_state on a channel to be 'closed'. This introduced race
conditions if packets were in flight from the peer. The recv_state
should only move into the closed state when a close is received from the
peer and delivered to the application. Calling abort or close should
only affect the send_state of the channel.