Avoid race in SSHChannel exit and exit_with_signal methods
This commit removes a check on send_state in the exit and
exit_with_signal functions in SSHChannel to avoid an exception being
raised when an applications calls these functions after the peer has
sent a close. AsyncSSH will attempt to send exit messages if a close
hasn't been sent in either direction at the time of the request, but
there's no way to guarantee that the exit message will be seen if
the peer initiates a close before the exit is received.