Catch OSError exception on write in SSHForwarder
This commit catches OSError exceptions that occur when forwarding data
in SSHForwarder, avoiding a traceback being printed about an uncaught
exception when debugging is enabled. The exception was being handled by
the connection_lost callback, but in some cases it would also throw the
exception in data_received or eof_received.