Fix clean up of channels getting an error during startup
This commit fixes a recently introduced problem that prevented channels
from being cleaned up properly when an error occurred during startup
(before a shell, exec, or subsystem request was processed). During
startup, data reading was paused, and that prevented the channel close
from being reported to the application, causing the channel to not be
cleaned up. Now, startup errors like this will immediately deliver the
close and the error which triggered it and the channel will be cleaned up.