Avoid sending responses to channel requests after close
This commit avoids sending responses to channel requests if the
processing of the request resulted in a channel close. No messages
should be sent on the channel after a close is sent.

This commit also switches from naked excepts to "except Exception"
to avoid reporting tracebacks on exceptions like GeneratorExit, and
disables branch coverage related to async actions completing after
being cancelled.