Allow channel requests when local side is closed
This commit allows channel requests to be received even after the local
sode of a channel is closed. This allows exit status and exit signals
to be recorded if they arrive between when the local side of a channel
is closed and when the remote close is received.

This commit also removes a state check when opening a channel which was
not needed since process_open is always called on a fresh channel
object.