Fix error reporting for close with incomplete Unicode data in buffer
This commit fixes reporting of errors when a channel is closed with
buffered data that doesn't make up a complete Unicode character.
Previously, this error was reported when EOF was received, but not if
the channel was closed without first sending an EOF.

This commit also simplifies some cleanup logic on channels and disables
code coverage for some defensive code that should never execute.