Rework fix for unicode error on connection close
This commit reworks the previous fix for not raising an error about
incomplete unicode on a connection close. Since the connection is
already closed, we shouldn't ever raise a DisconnectError at this point.
Attempting to do so results in an uncaught error in the cleanup callback
for the connection object.